fixed docker
